services/schedule/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
mysqlclient==2.2.7
gunicorn==20.1.0
python-dotenv==1.0.0